Reorganize components into domain-based subdirectories
- card/: Card, CardDetail, CardList
- deck/: Deck, DeckList, DeckImagePreview
- popup/: ExportPopup, ImportPopup
- ui/: SideMenu, TabButtons, ShareButtons, FireBaseLogin
- sections/: HeroSection, ExplanationSection, Footer (unchanged)

Fix 18 import paths across all files to reflect new locations
