+++ src/app/(app)/deck/2pick/components/DeckViewPopup.tsx
@@ -11,7 +11,7 @@
 
 import React from 'react';
 import { CardInfo } from '@/types/card';
-import Deck from '@/components/Deck';
+import Deck from '@/components/deck/Deck';
 import { css } from 'styled-system/css';
 import { button } from 'styled-system/recipes';
