+++ src/app/(app)/deck/[userId]/[deckId]/DeckPageClient.tsx
@@ -7,16 +7,16 @@ import { allYojoCards, allSweetCards, allPlayableCards } from '@/data/cards';
 import { doc, setDoc } from 'firebase/firestore';
 import { db } from '@/lib/firebase';
 import { validateDeckData } from '@/lib/schema';
-import CardList from '@/components/CardList';
+import CardList from '@/components/card/CardList';
 import { useAuth } from '@/lib/auth';
-import ExportPopup from '@/components/ExportPopup';
-import ImportPopup from '@/components/ImportPopup';
+import ExportPopup from '@/components/popup/ExportPopup';
+import ImportPopup from '@/components/popup/ImportPopup';
 import { GoogleAuthProvider, signInWithPopup } from 'firebase/auth';
 import { auth } from '@/lib/firebase';
-import ShareButtons from '@/components/ShareButtons';
+import ShareButtons from '@/components/ui/ShareButtons';
 import { useSettings } from '@/app/SideMenuProvider';
-import DeckList from '@/components/DeckList';
-import TabButtons, { TabDefinition } from '@/components/TabButtons';
+import DeckList from '@/components/deck/DeckList';
+import TabButtons, { TabDefinition } from '@/components/ui/TabButtons';
 import { css } from 'styled-system/css';
 
 const yojoLimit = 20;
