+++ src/components/card/CardDetail.tsx
