+++ src/components/deck/DeckImagePreview.tsx
